Move user-manager.service to services folder
