+++ src/app/app.module.ts
@@ -20,7 +20,7 @@ import {AutocompleteDropdownListItemComponent} from "./components/autocomplete-d
 import {AutocompleteDropdownComponent} from "./components/autocomplete-dropdown/autocomplete-dropdown.component";
 import {LanguageProxy} from "./services/proxy/language-proxy.service";
 import {LoginComponent} from "./views/login/login.component";
-import {UserManager} from "./user-manager.service";
+import {UserManager} from "./services/user-manager.service";
 import {JobCreateComponent} from "./views/jobs/job-create/job-create.component";
 import {JobPreviewComponent} from "./components/job-preview/job-preview.component";
 import {JobsComponent} from "./views/jobs/jobs.component";
