+++ src/app/views/home/home.component.ts
@@ -2,7 +2,7 @@ import {Component, OnInit} from "@angular/core";
 import {JobProxy} from "../../services/job-proxy.service";
 import {Job} from "../../models/job/job";
 import {TranslationService} from "../../services/translation.service";
-import {UserManager} from "../../user-manager.service";
+import {UserManager} from "../../services/user-manager.service";
 import {User} from "../../models/user";
 
 @Component({
