+++ src/app/views/job-details/job-details.component.ts
@@ -2,7 +2,7 @@ import {Component, OnInit} from "@angular/core";
 import {ActivatedRoute} from "@angular/router";
 import {Job} from "../../models/job/job";
 import {JobProxy} from "../../services/job-proxy.service";
-import {UserManager} from "../../user-manager.service";
+import {UserManager} from "../../services/user-manager.service";
 import {User} from "../../models/user";
 
 @Component({
